Bug 1309834. r=wchen. a=gchang
MozReview-Commit-ID: FO7bXlx6lv

--HG--
extra : source : 8704571a427f466c9143f46663862fb979429d7a

+++ parser/html/nsHtml5TreeBuilderHSupplement.h
@@ -138,7 +138,7 @@
      *                next call to this method.
      * @return true if successful; false if out of memory
      */
-    bool EnsureBufferSpace(size_t aLength);
+    bool EnsureBufferSpace(int32_t aLength);
 
     void EnableViewSource(nsHtml5Highlighter* aHighlighter);
